firewall-controller/requirements.txt
mysql-connector-python>=8.0,<9.0
python-dateutil>=2.8,<3.0
